# Conflicts:
#	docker-compose.yml
#	src/SimpleCliniq.Api/Controllers/Core/Diagnosa/DTDEndpoints.cs
#	src/SimpleCliniq.Api/Controllers/SampleDBContext.cs
#	src/SimpleCliniq.Api/Program.cs
#	src/SimpleCliniq.Api/Properties/launchSettings.json
#	src/SimpleCliniq.Api/SimpleCliniqApi.csproj
#	src/SimpleCliniq.Module.Core.Domain/Models/MDiagnosa.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MDokter.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MDtd.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MHargaRekanan.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MJadwalDokter.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MLaboratoriumHarga.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MLaboratoriumRekanan.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MMap.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MMorfologi.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MPasien.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MRadiologiHarga.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MRadiologiRekanan.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MRekanan.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MRuang.cs
#	src/SimpleCliniq.Module.Core.Domain/Models/MUser.cs
#	src/SimpleCliniq.Module.Core.Infrastructure/Database/Configuration/builder.cs

+++ src/SimpleCliniq.Module.Core.Domain/Models/MKamarinapHarga.cs
@@ -13,7 +13,7 @@ public partial class MKamarinapHarga
 
     public long? IdKamarInap { get; set; }
 
-    public decimal? RekananId { get; set; }
+    public int? RekananId { get; set; }
 
     public decimal? Harga { get; set; }
